+++ emav/probe_w.c
@@ -317,7 +317,7 @@ z_surge(float *local_zl, int na)
 
     z_chr(zc);
     local_z0 = zc[0];
-    R = (unsigned long)((double *) calloc(na * 2, sizeof(double)));
+    R = (unsigned long)((double *) calloc((size_t)(na * 2), sizeof(double)));
     c_ones(R, na);
     fd_window(R, na);
     w0 = mean_real(R, na);
@@ -361,14 +361,14 @@ thev_init(void)
 {
     int m;
 
-    qq1 = (float *) calloc(MAX_NF * 2 * 2, sizeof(float));
+    qq1 = (float *) calloc((size_t)(MAX_NF * 2 * 2), sizeof(float));
     if (qq1 == NULL) {
         decide(0, 1, "Can't allocate thev memory (qq1)");
         return (0);
     }
 
     os = MAX_NF * 2 * 3;	//allow for ps, zs, px
-    pzs = (unsigned long)((float *) calloc(os * 2, sizeof(float)));
+    pzs = (unsigned long)((float *) calloc((size_t)(os * 2), sizeof(float)));
     if (pzs == NULL) {
         decide(0, 1, "Can't allocate thev memory (pzs)");
 	free(qq1);
@@ -377,7 +377,7 @@ thev_init(void)
 
     m = (nthc < 6) ? 6 : nthc;	//allow for pl, zl, pf, z0, pr, cd
     ol = MAX_NF * 2 * m;;
-    pzl = (unsigned long)((float *) calloc(ol * 2, sizeof(float)));
+    pzl = (unsigned long)((float *) calloc((size_t)(ol * 2), sizeof(float)));
     if (pzl == NULL) {
         decide(0, 1, "Can't allocate thev memory (pzl)");
 	free(qq1);
@@ -1220,7 +1220,7 @@ cavity_length(float *local_pc, int nf, double vs)
     int i, ii, ir, i1, i2, m, n;
 
     n = nf * 2 + 2;
-    p = (unsigned long)((float *) calloc(n, sizeof(float)));
+    p = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
     for (i = 0; i < nf; i++) {
 	ir = 2 * i;
 	ii = 2 * i + 1;
@@ -1743,9 +1743,9 @@ thev_adjust(char *cfn, char *sfn, SAV_FFT *pft, int sil)
 	    free_null(iadj);
 	    free_null(refl);
 	    n = (nthf < pft[j].npts) ? nthf : pft[j].npts;
-	    padj = (unsigned long)((float *) calloc(n, sizeof(float)));
-	    iadj = (unsigned long)((float *) calloc(n, sizeof(float)));
-	    refl = (unsigned long)((float *) calloc(n, sizeof(float)));
+	    padj = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
+	    iadj = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
+	    refl = (unsigned long)((float *) calloc((size_t)(n), sizeof(float)));
 	    for (i = 0; i < n; i++) {
 		ir = 2 * i;
 		ii = 2 * i + 1;
